+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 ### New in 0.72 (not released yet)
 
+* New: Configure JSON serialization: 
+  ```csharp
+  EventFlowOptions.New.
+    .ConfigureJson(json => json
+      .AddSingleValueObjects()
+      .AddConverter<SomeConverter>()
+    )
+  ```
 * New: `EventFlow.TestHelpers` are now released as .NET Standard as well
 * Fix: Upgrade `EventStore.Client` to v5.0.1 and use it for both .NET Framework and .NET Core
 * Fix: Storing events in MS SQL Server using `MsSqlEventPersistence` now correctly